+++ libs/services/ListService.php
@@ -1,6 +1,4 @@
 <?php
-require_once "ArticleService.php";
-
 class AgreableListService {
   protected static $dupes = array();
