+++ single.php
@@ -9,7 +9,7 @@
  * @since    Timber 0.1
  */
 
-require_once "libs/services/ArticleService.php";
+require_once "libs/services/WidgetService.php";
 require_once "libs/services/CategoryService.php";
 
 $context = Timber::get_context();
